ratatouille_app/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
deprecation==2.1.0
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
Faker==25.0.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.15.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gotrue==2.4.2
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.24.0
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packaging==24.0
parso==0.8.4
postgrest==0.16.4
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
StrEnum==0.4.15
text-unidecode==1.3
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
WTForms==3.1.2
